+++ src/main/java/uk/gov/cslearning/record/domain/Event.java
@@ -7,22 +7,22 @@ public class Event {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(nullable = false)
     private String path;
 
     public Event() {}
 
-    public Event(Long id){
+    public Event(int id){
         this.id = id;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
